Adding Oscilloscope.
This is a first implementation of the oscilloscope. It’s actually just
a live wave form viewer right now and doesn’t have any real features of
an oscilloscope.

The implementation is based on UDP sending of audio data from a fixed
port in Moodler to a fixed port in the Oscilloscope. We could change
that to not require that the oscilloscope has a fixed port but that’s
not too important for the first working version.

For the graphics it uses GLFW and OpenGL; a very simple vertex and
fragment shader to draw the curves. I’m sure I’m doing something wrong
..