Fix overflow accumulation.
I have a fix for the overflow issues we were seeing and the
accumulation of the voices is now done as doubles for maximum audio
fidelity.  We only convert to short at the last minute.

This appears to clear up the issues in the previous version where we
were overflowing short and wrapping around for some reason.